model/requirements.txt
Flask
gunicorn
google-cloud-storage
pillow
numpy
tensorflow==2.13.0
keras
google-cloud-firestore